Fixed type used for number

+++ vertigo-orchestra-api/src/main/java/io/vertigo/orchestra/definition/ProcessDefinition.java
@@ -28,7 +28,7 @@ public interface ProcessDefinition {
 
 	List<ActivityDefinition> getActivities();
 
-	Long getRescuePeriod();
+	int getRescuePeriod();
 
 	Option<String> getMetadatas();
